refactor: use flashinfer Python API directly, remove C++ attention ops bridge
- Delete xllm_attention_ops.cpp/h (batch_decode/prefill/update_plan torch ops
  were only used by the Python executor, now replaced by flashinfer Python API)
- Move reshape_paged_cache registration into xllm_ops_library.cpp
- layers/attention.py: use flashinfer BatchDecodeWithPagedKVCacheWrapper /
  BatchPrefillWithRaggedKVCacheWrapper directly instead of torch.ops.xllm_ops
- Rename py_model_bridge → py_model_helper, consolidate PyDictVisitor and
  dtype_to_string utilities
- Move py_causal_lm to models/llm/ (where other CausalLM implementations live)
- Add core/kernels/xllm_torch_ops.h as device-agnostic dispatch header
- Remove USE_CUDA guard from PyCausalLM compilation (device-agnostic)
- Add FlashinferWorkspace skip for model_impl=python in llm_worker_impl.cpp
