--- xllm/core/kernels/cuda/xllm_attention_ops.h
@@ -1,5 +1,0 @@
-/* Copyright 2025-2026 The xLLM Authors. */
-#pragma once
-namespace xllm {
-void ensure_xllm_attention_ops_registered();
-}
